Mostly function v2
Missing some features, but is able to create usable compilations
